+++ browser_tests/colorPalette.spec.ts
@@ -1,6 +1,6 @@
 import { expect } from '@playwright/test'
 
-import type { Palette } from '../src/types/colorPaletteTypes'
+import type { Palette } from '../src/types/colorPaletteSchema'
 import { comfyPageFixture as test } from './fixtures/ComfyPage'
 
 const customColorPalettes: Record<string, Palette> = {
