+++ src/schemas/colorPaletteSchema.ts
