+++ src/schemas/keybindingSchema.ts
