+++ src/services/litegraphService.ts
@@ -13,7 +13,7 @@ import { IBaseWidget, IWidget } from '@comfyorg/litegraph/dist/types/widgets'
 import { useNodeImage, useNodeVideo } from '@/composables/node/useNodeImage'
 import { st } from '@/i18n'
 import { ComfyNodeDef } from '@/schemas/apiSchema'
-import type { NodeId } from '@/schemas/comfyWorkflow'
+import type { NodeId } from '@/schemas/comfyWorkflowSchema'
 import { ANIM_PREVIEW_WIDGET, ComfyApp, app } from '@/scripts/app'
 import { $el } from '@/scripts/ui'
 import { calculateImageGrid, createImageHost } from '@/scripts/ui/imagePreview'
