+++ src/stores/workflowStore.ts
@@ -2,7 +2,7 @@ import _ from 'lodash'
 import { defineStore } from 'pinia'
 import { computed, markRaw, ref } from 'vue'
 
-import { ComfyWorkflowJSON } from '@/schemas/comfyWorkflow'
+import { ComfyWorkflowJSON } from '@/schemas/comfyWorkflowSchema'
 import { api } from '@/scripts/api'
 import { ChangeTracker } from '@/scripts/changeTracker'
 import { defaultGraphJSON } from '@/scripts/defaultGraph'
