+++ src/types/comfy.ts
@@ -2,8 +2,8 @@ import type { LGraphNode } from '@comfyorg/litegraph'
 import { Positionable } from '@comfyorg/litegraph/dist/interfaces'
 
 import type { ComfyNodeDef } from '@/schemas/apiSchema'
-import type { ComfyWorkflowJSON } from '@/schemas/comfyWorkflow'
-import type { Keybinding } from '@/schemas/keyBindingTypes'
+import type { ComfyWorkflowJSON } from '@/schemas/comfyWorkflowSchema'
+import type { Keybinding } from '@/schemas/keyBindingSchema'
 import type { ComfyApp } from '@/scripts/app'
 import type { ComfyWidgetConstructor } from '@/scripts/widgets'
 import type { ComfyCommand } from '@/stores/commandStore'
