+++ src/types/litegraph-augmentation.d.ts
@@ -4,7 +4,7 @@ import type { LLink, Size } from '@comfyorg/litegraph'
 import type { ComfyNodeDef } from '@/schemas/apiSchema'
 import type { DOMWidget, DOMWidgetOptions } from '@/scripts/domWidget'
 
-import type { NodeId } from '../schemas/comfyWorkflow'
+import type { NodeId } from '../schemas/comfyWorkflowSchema'
 
 /** ComfyUI extensions of litegraph */
 declare module '@comfyorg/litegraph/dist/types/widgets' {
